MOONWLKR REDESIGN PRODUCT PAGE SCRIPTS These scripts relate to the product display section only
This file contains the following scripts:
1. Create a quantity selector with '+' and '-' buttons instead of the default input number selector
2. Add variant tooltips indicating discounts to products with 2 or 3 variants. The tooltip appears above the variant button.
3. Dinamically update the "Add to Cart" button text according to product availability. If out of stock it will change to "out of stock".
4. Attempt at dinamically updating the product price on the front-end according to selected quantity and variant. The product regular and sale price values are taken from the price divi block if the product is single variant, or from the variant product element inside the cart form. Variant prices are hidden as to not show two price elements but their values can still be retrieved.
The goal is to have a single script to update the product price as to not repeat code -functions, variables- for every possible scenario.
Currently it's not working for all scenarios, and I believe it's due to issues with comparison logic and with updating variable values. It works fine for products that are both on sale and have multiple quantity variants. It's not working if the product is single variant and/or is not on sale. Will keep working on this and push with updates until it's working and live on prod.