Layed out first approach at calculating and updating prices.
Most products are always on sale so this first approach reflects this scenario.

Current price (which in this scenario is sale price) can be grabbed from a global variable called "item". "item.price" will give us that value. It's important to grab this value from this source instead of grabbing it from the sale price HTML element, as that will change every time the price is updated by increasing or decreasing the desired quantity.

The regular price however has to be taken from its HTML element when the item is on sale, as the item variable only contains the current price. It's not ideal because this HTML element will be updated to reflect a new price value as customers increase or decrease the desired quantity. It's then important to grab it immediately on page load to make sure we have that default regular price value stored in a variable.

There are two functions to calculate each price, which will then be used in another function (updatePrice). updatePrice() will grab the calculated prices and update both sale and regular prices' HTML elements to reflect them.

This logic may change as I adapt it to match all possible price scenarios.