Added base for another scenario: if product is not on sale and does not have variants
This means that the regular price can be retrieved from the item variable "item.price", and that the regular price HTML element isn't separated into two (regular <del> and sale <ins>).

Added condition to check if the product is on sale on the calculateRegularPrice function, in order to know where to grab the regular price value from and attribute it to the regularPrice variable.

Added condition to check if the product is on sale on the updatePrice function and moved its original logic inside it. Then wrote logic to update regular price if product is not on sale.