Added logic to update price after clicking in a product variant.
Currently it's not dependant on the calculate prices functions or the update price function.

After clicking on a variant, I have to retrieve the regular and sale prices from the variant form JSON (shoutout to Luis for the tip!). The variant form JSON retrieves as many objects as there are variations. So to access the correct object price values, I have to store the clicked button index in a variable, then pass it into the JSON.parse .

Then it calculates the prices by multiplying them with the quantity value.

Then it replaces the price containers' inner HTML with the calculated prices' values.

My next step is to refactor this to see if it's possible to include it in the previously coded functions (maybe pass the clicked button index as an argument to correctly grab the clicked variation prices).