Changes to calculate and update pricing functions to include product variations
Created the calculatePrice function. It multiplies the quantity value by a parameter value.

This parameter will be a regular price or a sale price according to these different scenarios:
1. Product has no variations and is not on sale - only requires the regular price;
2. Product has no variations but is on sale - requires both regular and sale prices;
3. Product has variations and all are on sale - requires both regular and sale prices but they're taken from the variations form JSON objects;

The parameter is passed on the price function that includes the scenarios above.

For the last scenario I created a variable to contain the index of the clicked variation button, as a way to then access that same index on the variation form json objects. The default value is 0 because that's the default variation on page load.

On the variation forEach click event listener, the index variable is updated with the index of the button that was just clicked. Then it runs the price() function and the updatePriceElements() function.

My next step is to test this on moonwlkr dev.