Created new branch to approach an issue with the Amanita products price calculation
The 4count variation selected by default has a max quantity of 1 ('max' attribute on the input element), while the 10count variation doesn't. What happens is that if I change the quantity while on the 10count variation, the price updates fine but if I then click on the 4count variation, it calculates the price with the same quantity as 10count, but showing 1 on the input.

All I did was create an if statement inside the variationButtons forEach click event, where it checks if there's a max quantity of 1. If true the inputQtyNumber variable value is changed to 1, which will then be taken into account on the calculatePrice function.  It's basically a quantity value reset.

Also had to wrap all of that in a setTimeout because it takes a little bit for the input to change the max value.

Tested this on moonwlkrdev and it's working!