Created this branch to create dynamic creation of discount tooltips. These tooltips are supposed to appear above the variation buttons (when product has variations).
What I idealized before was a function to create each of the tooltips given that each would have different text on them (if 1 variation only 10% off; if 2 variations, 2nd is 15% off);

But both functions had similar processes so I worked to find a way to create a single function to dynamically add these tooltips depending on the amount of variations.

I'll explain in more detail once I comment this function on the file.