Code changes after James reviewed it - Thank you again James!!!
Set the whole script as a callback function to run on setTimeout for readibility as James suggested. Changed the name of the function as it contains more than the dynamic pricing script.

Also changed variationButtons querySelectorAll to getElementsByClassName as per James, but had to transform the nodeList into an array to run the necessary forEach functions. Took this opportunity to also review the createTooltips function as now the variationButtons variable is an array.

Testing the new script on dev I noticed that the "out of stock" one was not working anymore. For some reason, the add to cart button now takes longer to load the class that disables it when the product is out of stock. So I moved this code into a function running with another setTimeout function with a delay of 500.

Updated product-page-scripts-old.js to hold the current script running and working on prod, so that we can roll back to it if there are any issues when implementing all the new scripts.

Removed the dynamic-pricing.js file from the main branch as it is no longer necessary. I'm planning to add it again once I've reorganized this repository for Moonwlkr redesign.

If upon review this is all working, I think we can move this to prod!